создаёт и подключает модуль randomizers
создаёт и подключает модуль для функции получения случайного числа и функции получения случайного элемента
