+++ js/main.js
@@ -1,15 +1,5 @@
 import {PHOTOS_URL, PHOTOS_DESCRIPTION, AVATARS_URL, MESSAGES, NAMES} from './data.js';
-
-// Функция для генерирования случайного числа в заданном диапазоне из проекта "Код и Магия"
-const getRandomInteger = (a, b) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
-
-// Функция для получения случайного элемента из проекта "Код и Магия"
-const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
+import {getRandomInteger, getRandomArrayElement} from './randomizers.js';
 
 // Функция вызвращает случайный комментарий содержащий один или два сообщения. Количество сообщений определяется случайным образом
 const getRandomMessage = function () {
