+++ js/randomizers.js
@@ -1,0 +1,12 @@
+// Функция для генерирования случайного числа в заданном диапазоне из проекта "Код и Магия"
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+// Функция для получения случайного элемента из проекта "Код и Магия"
+const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
+
+export {getRandomInteger, getRandomArrayElement};
